+++ did/did-agent-storage/lib/index.js
@@ -5,6 +5,7 @@ const { EventEmitter } = require('events');
 /**
  * Defines the interface of authorized did auth storage
  * Which is used when building an centralized service that do did-auth on behalf of many applications
+ * The service can implement a version of itself
  *
  * @class DidAgentStorage
  * @see @arcblock/did-agent-storage-memory